% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{border_coord_df}
\alias{border_coord_df}
\title{Continental U.S. border coordinates.}
\format{A data frame with 21,603 rows and 4 variables:
\describe{
  \item{lat}{Latitude of mile-long border point.}
  \item{long}{Longitude of mile-long border point.}
  \item{bordindx}{Unique border-segment (e.g., AL-FL) numeric identifer.}
  \item{st1st2}{Unique alphabetized border-segment (e.g., AL-FL) identifier.}
}}
\source{
\url{http://users.econ.umn.edu/~holmes/data/BorderData.html}
}
\usage{
border_coord_df
}
\description{
A dataset containing coordinates for mile-long strips of all borders between states in
the continental U.S.
}
\keyword{datasets}
