% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bscp_df}
\alias{bscp_df}
\title{Border segment county assignments.}
\format{A data frame with 1,184 rows and 5 variables:
\describe{
  \item{fips_code}{County FIPS code, in original string format. Unique identifier.}
  \item{county_state}{State postal code corresponding to FIPS code.}
  \item{dist_to_border}{Distance (in miles) of population center to the closest adjacent
  state border strip.}
  \item{dist_to_segment}{Distance (in miles) to closest 50 mile border segment.}
  \item{bscp_id}{50-mile long border segment identifier.}
}}
\source{
See \code{adjacent_county_df}.
}
\usage{
bscp_df
}
\description{
A dataset that assigns each border county to the closest 50 mile-long border strip. A total of
73 counties are the only county assigned to a border strip; these counties will be ignored
in any panel estimates. Total of 421 unique border segment identifiers with counties.
Distribution of border segments by number of asssigned counties: 73 with
1 county, 124 with 2 counties, 97 with 3 counties, 83 with 4 counties, 34 with 5 counties, 4
with 6 counties, 3 with 7 counties, 2 with 8 counties, 1 with 9 counties. Note that only segments
with counties in multiple states will have variation exploited. This breakdown is: 95 segments
with 1 state, 311 segments with 2 stats, and 15 segments with 3 states.
}
\keyword{datasets}
