% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbzip_df}
\alias{cbzip_df}
\title{ZIP code border assignments.}
\format{A data frame with 38,879 rows and 14 variables:
\describe{
  \item{zip_code}{ZIP code, in original string format. Unique identifier.}
  \item{fips_code}{County FIPS code, in original string format.}
  \item{state}{State postal code corresponding to FIPS code.}
  \item{dist_to_border}{Distance to closest state border, in miles. Null if no ZCTA match.}
  \item{nearest_index}{Nearest border index. Null if no ZCTA match.}
  \item{nearest_border}{Nearest state border. Null if no ZCTA match.}
  \item{dist_to_segment_id}{Distance to closest 20-mile border segment. Null if no ZCTA match.}
  \item{cb_segment_id}{Nearest 20-mile border segment. Null if no ZCTA match.}
  \item{cpcp_id}{Couplet cluster pair identifier for county. Null if not border county or
  if there is no couplet identifier match.}
  \item{relaxed_cpcp_id}{Relaxed couplet cluster pair identifier for county. Null if not border
  county or if there is no relaxed couplet identifier match.}
  \item{mxcp_id}{Max-method cluster identifier for county. Null if not border county or if there
  is no max-method match.}
  \item{relaxed_mxcp_id}{Relaxed max-method cluster identifier for county. Null if not border
  county or if there is no relaxed max-method match.}
  \item{border_county_flag}{Binary flag for whether ZIP resides in a border county.}
}}
\source{
See \code{zip_df}.
}
\usage{
cbzip_df
}
\description{
A dataset that facilitates the use of ZIP codes in cross-border analyses. Contains all ZIP
codes in the continental United States (i.e., there is no filtering). Provides the following
identifiers for conducting cross-border analyses: dist_to_border, nearest_borer,
dist_to_segment_id, cb_segment_id, state_border_id, cpcp_id, relaxed_cpcp_id, mxcp_id,
relaxed_mxcp_id, border_county_flag. Includes ZIP codes in and not in border counties, far and
close to state borders or 20 mile long border segments, and those that can and cannot be
mapped to a ZCTA centroid.
}
\keyword{datasets}
