% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_df}
\alias{census_df}
\title{U.S. states}
\format{A data frame with 51 rows and 5 variables:
\describe{
  \item{census_region}{Census region of state.}
  \item{census_division}{Census division of state.}
  \item{state_fips}{Two-digit FIPS code for state.}
  \item{state_code}{Post service state code.}
  \item{state_name}{Name of state.}
}}
\usage{
census_df
}
\description{
A dataset containing census region and division identifiers for the 50 U.S. states and D.C.
}
\keyword{datasets}
