% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{county_dist}
\alias{county_dist}
\title{Distance between county population centers.}
\usage{
county_dist(fips_one, fips_two, miles = TRUE)
}
\arguments{
\item{fips_one}{FIPS code of first county. Order does not matter.}

\item{fips_two}{FIPS code of second county. Order does not matter.}

\item{miles}{Binary flag for whether to return distance in miles (TRUE) or kilometers (FALSE).}
}
\description{
\code{county_dist} measures the geographic distance between two counties. The location of each
county is assumed to be its center of population. Function is vectorized and accounts for
counties with NA matches to geographic coordinate data.
}
\examples{
county_dist(fips_one = "01003", fips_two = "01005", miles = TRUE)
county_dist(fips_one = temp_df$fips_code, fips_two = temp_df$neighbor_fips_code)
}
