% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{county_to_state_border}
\alias{county_to_state_border}
\title{Distance between county population center and nearest state border.}
\usage{
county_to_state_border(fips_code, return_state_border_id = FALSE,
  miles = TRUE)
}
\arguments{
\item{fips_code}{FIPS code of county.}

\item{return_state_border_id}{Binary flag for whether to return the state border identifier
for the nearest state border instead of the distance. Defaults to FALSE.}

\item{miles}{Binary flag for whether to return distance in miles (TRUE) or kilometers (FALSE).}
}
\description{
\code{county_to_state_border} outputs the minimum distance between a county population center
and the nearest state border. Alternatively, returns the identifier for the nearest state border.
}
\examples{
county_to_state_border(fips_one = "01003", return_state_border_id = FALSE, miles = TRUE)
county_to_state_border(fips_one = temp_df$fips_code, return_state_border_id = TRUE)
}
