% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpcp_df}
\alias{cpcp_df}
\title{Couplet and relaxed couplet county assignments.}
\format{A data frame with 1,184 rows and 6 variables:
\describe{
  \item{fips_code}{County FIPS code, in original string format. Unique identifier.}
  \item{county_state}{State postal code corresponding to FIPS code.}
  \item{cpcp_id}{Couplet cluster pair identifier.}
  \item{relaxed_cpcp_id}{Relaxed couplet cluster pair identifier.}
  \item{cpcp_remove_flag}{Binary flag equal to one if a county is unmatched to a couplet
  cluster.}
  \item{relaxed_cpcp_remove_flag}{Binary flag equal to one if a county is unmatched to a
  relaxed couplet cluster.}
}}
\source{
See \code{adjacent_county_df}.
}
\usage{
cpcp_df
}
\description{
A dataset that assigns each county to a unique cluster using the "couplet" algorithm. Couplet
clusters are defined as cross-border county pairs in which each constituent county does not
appear in any other cross-border county pairs. Because each border county may belong to several
cross-border county pairs, the "couplet" algorithm is needed to partition the space with the
goal of maximizing the total number of couplet clusters. Couplet clusters are identified by
\code{cpcp_id}; some counties are excluded from any couplet clusters by the algorithm. To
resolve the issue of non-assignment of some border counties, the \code{relaxed_cpcp_id}
identifier assigns non-matched border counties to the couplet their cross-borde neighbor.
Metadata includes: (1) 499 unique couplet clusters mapped to 998 border counties, (2) 186
border counties without a matched couplet cluster, (3) 499 unique relaxed couplet clusters
mapped to 1,183 border counties, (4) 1 border county without a matched relaxed couplet cluster.
}
\keyword{datasets}
