% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sbscp_df}
\alias{sbscp_df}
\title{State border strip county assignments.}
\format{A data frame with 1,184 rows and 3 variables:
\describe{
  \item{fips_code}{County FIPS code, in original string format. Unique identifier.}
  \item{county_state}{State postal code corresponding to FIPS code.}
  \item{state_border_id}{Border identifier for closest state border strip.}
  \item{dist_to_border}{Distance (in miles) of population center to the closest adjacent
  state border strip.}
  \item{num_counties_in_strip}{Number of unique adjacent counties belong to state border pair
  strip. Fixed effect specifications will eliminate those with a single county.}
  \item{num_states_in_strip}{Number of unique states corresponding to unique adjacent counties
  that belong to the state border pair strip. Fixed effect specifications focused on cross-
  border variation will eliminate those with a single state in the state border pair.}
}}
\source{
See \code{adjacent_county_df}.
}
\usage{
sbscp_df
}
\description{
A dataset that assigns each border county to a unique state border pair strip (e.g., AL-FL).
For counties that reside along multiple state-border strips (e.g., AZ-NV, NV-UT), the county
is assigned to the state-border strip to which its center of population is closest. If a county
is closest to a state-border strip that it is not adjacent with, the next closest adjacent
state-border strip is chosen. State-border pair strips are identified by \code{state_border_id}.
}
\keyword{datasets}
