% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_df}
\alias{state_df}
\title{U.S. states}
\format{A data frame with 51 rows and 2 variables:
\describe{
  \item{state}{Name of state in capital letters.}
  \item{state_code}{Postal service state code.}
}}
\usage{
state_df
}
\description{
A dataset containing identifiers for the 50 U.S. states and D.C.
}
\keyword{datasets}
