% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{zip_dist}
\alias{zip_dist}
\title{Distance between ZCTA centroids.}
\usage{
zip_dist(zip_one, zip_two, miles = TRUE)
}
\arguments{
\item{miles}{Binary flag for whether to return distance in miles (TRUE) or kilometers (FALSE).}

\item{fips_one}{ZIP code of first ZCTA. Order does not matter.}

\item{fips_two}{ZIP code of second ZCTA. Order does not matter.}
}
\description{
\code{zip_dist} measures the geographic distance between two ZCTAs. The location of each
ZCTA is assumed to be its centroid. Returns NA values for ZIP codes that do not match to a
ZCTA. Function is vectorized and accounts for NA matches.
}
\examples{
zip_dist(fips_one = "89128", fips_two = "89034", miles = TRUE)
zip_dist(fips_one = temp_df$zips_one, fips_two = temp_df$neighbor_zips)
}
