% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{zip_to_state_border}
\alias{zip_to_state_border}
\title{Distance between ZCTA centroids and nearest state border.}
\usage{
zip_to_state_border(zip_code, return_state_border_id = FALSE,
  miles = TRUE)
}
\arguments{
\item{zip_code}{Five-digit ZIP code of ZCTA}

\item{return_state_border_id}{Binary flag for whether to return the state border identifier
for the nearest state border instead of the distance. Defaults to FALSE.}

\item{miles}{Binary flag for whether to return distance in miles (TRUE) or kilometers (FALSE).}
}
\description{
\code{zip_to_state_border} outputs the minimum distance between a ZCTA centroid and the
nearest state border. Alternatively, returns the identifier for the nearest state border.
}
\examples{
zip_to_state_border(fips_one = "89128", return_state_border_id = FALSE, miles = TRUE)
zip_to_state_border(fips_one = temp_df$zip_code, return_state_border_id = TRUE)
}
